Completed Permenant Delete Posts Functionality

+++ app/Http/Controllers/PostsController.php
@@ -3,6 +3,8 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Http\Requests\Posts\CreatePostsRequest;
 use App\Post;
+use Illuminate\Support\Facades\Storage;
+
 class PostsController extends Controller
 {
     /**
@@ -82,12 +84,26 @@ class PostsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Post $post)
+    public function destroy($id)
     {
-        $post->delete();
-
-        session()->flash('success', 'Post trashed successfully.');
-
+        $post = Post::withTrashed()->where('id', $id)->firstOrFail();
+        if ($post->trashed()) {
+            Storage::delete($post->image);
+            $post->forceDelete();
+          } else {
+            $post->delete();
+          }
+        session()->flash('success', 'Post deleted successfully.');
         return redirect(route('posts.index'));
+    }
+    /**
+     * Display a list of all trashed posts
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function trashed()
+    {
+      $trashed = Post::withTrashed()->get();
+      return view('posts.index')->with('posts', $trashed);
     }
 }
